+++ Testdb/src/main/java/com/example/demo/services/CardService.java
@@ -24,7 +24,7 @@ public class CardService {
 	public Iterable<Card> getAllByOwner_Id(int id) {
 		return cardrepo.findAllByOwner(id);
 	}
-	public Iterable<Card> getActiveByOwner_Id(int id) {
+	public Iterable<Card> getActiveByOwner_Id(Long id) {
 		return cardrepo.findActiveByOwner(id);
 	}
